Moved tool_call_utils into test_helpers package; asserted on tool call read results
