+++ tests/test_helpers/tool_call_utils.py
